# open-api/openapi.yaml
openapi: 3.0.3

# ---------------------------------------------------------------------------------
# Este archivo declara SOLO las rutas: que verbo, en que path, con que tag, y a que
# schema apunta cada peticion y cada respuesta.
#
# Ningun schema se define aqui. Todo se referencia con $ref hacia los otros tres
# archivos, para que este quede legible como indice de la API:
#
#   requests.yaml    lo que entra
#   resources.yaml   lo que sale (y los enums, que comparten ambos lados)
#   responses.yaml   el envoltorio HTTP de cada respuesta
#
# El `tags` de cada operacion decide el nombre de la interfaz Java generada:
# tag `cotizacion` -> CotizacionApi / CotizacionApiDelegate. Por eso los tags
# son singulares y en minusculas.
# ---------------------------------------------------------------------------------

info:
  title: API de Cotizacion de Envios
  description: |
    Ejercicio de contrato-primero. El contrato manda: los controllers y los modelos
    los genera openapi-generator, no se escriben a mano.
  version: 0.0.1

servers:
  - url: 'http://localhost:{port}'
    description: local
    variables:
      port:
        default: '8080'

tags:
  - name: cotizacion
    description: Cotizacion de envios y catalogo de zonas
  - name: propina
    description: Ejemplo de referencia, calculadora de propinas

paths:

  # =================================================================================
  # TU API  ·  paquete com.teletubies.endpoints.cotizacion
  # =================================================================================

  '/api/v1/cotizaciones':
    post:
      tags:
        - cotizacion
      summary: Cotiza un envio
      description: |
        Recibe origen, destino y peso del paquete, y devuelve el costo estimado junto
        con el tiempo de entrega.
      operationId: cotizarEnvio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './requests.yaml#/components/schemas/CotizarEnvioRequest'
      responses:
        "200":
          $ref: './responses.yaml#/components/responses/CotizacionResponse'
        "400":
          $ref: './responses.yaml#/components/responses/BadRequest'
        # -------------------------------------------------------------------------
        # TE TOCA A TI
        #
        # Falta la respuesta de regla de negocio incumplida: destino no soportado, o
        # paquete por encima del maximo. Ese caso NO es un 400 —el request esta bien
        # formado y se entendio— y hoy este contrato no lo declara.
        #
        # 1. Declara la response en responses.yaml (mira UnprocessableEntity ahi).
        # 2. Referenciala aqui abajo, con el codigo que decidas.
        # 3. Regenera:  make openapi-generate
        # 4. Manejala en tu @RestControllerAdvice.
        #
        # Documenta en el README que codigo elegiste y por que.
        # -------------------------------------------------------------------------
        "500":
          $ref: './responses.yaml#/components/responses/InternalServerError'

  '/api/v1/zonas':
    get:
      tags:
        - cotizacion
      summary: Lista las zonas de destino soportadas
      description: |
        Permite al cliente conocer los destinos validos ANTES de intentar cotizar.
      operationId: listarZonas
      responses:
        "200":
          $ref: './responses.yaml#/components/responses/ZonasResponse'
        "500":
          $ref: './responses.yaml#/components/responses/InternalServerError'

  # =================================================================================
  # EJEMPLO DE REFERENCIA  ·  paquete com.teletubies.endpoints.ejemplo
  #
  # Misma forma que tu API (un POST que calcula y un GET de catalogo) pero de otro
  # dominio. Leelo, y luego borra estas dos rutas junto con el paquete `ejemplo`.
  # =================================================================================

  '/api/v1/ejemplo/propinas':
    post:
      tags:
        - propina
      summary: Calcula la propina de una cuenta
      operationId: calcularPropina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './requests.yaml#/components/schemas/CalcularPropinaRequest'
      responses:
        "200":
          $ref: './responses.yaml#/components/responses/PropinaResponse'
        "400":
          $ref: './responses.yaml#/components/responses/BadRequest'
        "422":
          $ref: './responses.yaml#/components/responses/UnprocessableEntity'
        "500":
          $ref: './responses.yaml#/components/responses/InternalServerError'

  '/api/v1/ejemplo/monedas':
    get:
      tags:
        - propina
      summary: Lista las monedas soportadas
      operationId: listarMonedas
      responses:
        "200":
          $ref: './responses.yaml#/components/responses/MonedasResponse'
        "500":
          $ref: './responses.yaml#/components/responses/InternalServerError'